Спринт №10 refactor: добавление Optional

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -2,7 +2,9 @@ package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 
@@ -26,49 +28,69 @@ public class FilmController {
     public List<Film> findAll(HttpServletRequest request) {
         log.info("Получен запрос к эндпоинту: '{} {}', Строка параметров запроса: '{}'",
                 request.getMethod(), request.getRequestURI(), request.getQueryString());
-        return filmService.findAll();
+        return filmService.getAll();
     }
 
     @PostMapping(value = "/films")
     public Film create(@Valid @RequestBody Film film, HttpServletRequest request) {
         log.info("Получен запрос к эндпоинту: '{} {}', Строка параметров запроса: '{}'",
                 request.getMethod(), request.getRequestURI(), request.getQueryString());
-        return filmService.createFilm(film);
+        return filmService.create(film);
     }
 
     @PutMapping(value = "/films")
     public Film update(@Valid @RequestBody Film film, HttpServletRequest request) {
         log.info("Получен запрос к эндпоинту: '{} {}', Строка параметров запроса: '{}'",
                 request.getMethod(), request.getRequestURI(), request.getQueryString());
-        return filmService.updateFilm(film);
+        return filmService.update(film);
     }
 
     @GetMapping("/films/{id}")
     public Film getFilm(@PathVariable int id, HttpServletRequest request){
         log.info("Получен запрос к эндпоинту: '{} {}', Строка параметров запроса: '{}'",
                 request.getMethod(), request.getRequestURI(), request.getQueryString());
-        return filmService.findFilmById(id);
+        if (id > 0) {
+            return filmService.getById(id);
+        } else {
+            log.info("Некорректный запрос. Id должен быть больше 0");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Некорректный запрос. Id должен быть больше 0");
+        }
     }
 
     @PutMapping("/films/{id}/like/{userId}")
     public void addLike(@PathVariable Integer id, @PathVariable Integer userId, HttpServletRequest request) {
         log.info("Получен запрос к эндпоинту: '{} {}', Строка параметров запроса: '{}'",
                 request.getMethod(), request.getRequestURI(), request.getQueryString());
-        filmService.addLike(id, userId);
+        if ((id > 0) && (userId > 0)) {
+            filmService.addLike(id, userId);
+        } else {
+            log.info("Некорректный запрос. Id должен быть больше 0");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Некорректный запрос. Id должен быть больше 0");
+        }
     }
 
     @DeleteMapping("/films/{id}/like/{userId}")
     public void deleteLike(@PathVariable Integer id, @PathVariable Integer userId, HttpServletRequest request) {
         log.info("Получен запрос к эндпоинту: '{} {}', Строка параметров запроса: '{}'",
                 request.getMethod(), request.getRequestURI(), request.getQueryString());
-        filmService.removeLike(id, userId);
+        if ((id > 0) && (userId > 0)) {
+            filmService.removeLike(id, userId);
+        } else {
+            log.info("Некорректный запрос. Id должен быть больше 0");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Некорректный запрос. Id должен быть больше 0");
+        }
     }
 
     @GetMapping("/films/popular")
     public List<Film> getPopularFilms(@RequestParam(defaultValue = "10") Integer count, HttpServletRequest request) {
         log.info("Получен запрос к эндпоинту: '{} {}', Строка параметров запроса: '{}'",
                 request.getMethod(), request.getRequestURI(), request.getQueryString());
-        return filmService.findMostLikesFilm(count);
+        if (count > 0) {
+            return filmService.getMostLikesFilm(count);
+        } else {
+            log.info("Некорректный запрос. Количество должно быть больше 0");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Некорректный запрос. Количество должно быть больше 0");
+        }
     }
 
 }
